churn-prediction-dashboard/backend/requirements.txt
Django==5.2.5
djangorestframework==3.14.0
django-cors-headers==3.14.0
pandas==1.5.3
scikit-learn==1.2.2
numpy==1.24.2
gunicorn==21.2.0
whitenoise==6.4.0
python-dotenv==1.0.0